Use relative paths for configuration and output (#15)
The CLI has been modified slightly to look for the configuration in the
current working directory and not in `.github/`. This ensures that both
the configuration path as well as the path to the output file by default
use the current working directory, which is a more consistent and
predictable user experience.