Simplify groups in configuration
The configuration has been slightly simplified by changing groups from a
map to a list. Before, each group had an identifier in the form of
`[labels.<identifier>]`. The identifier was never used for anything, so
we can remove it and simplify the configuration to just use `[[group]]`
to collect them into a list.