Allow labels to be renamed via optional aliases
When a label needs to be renamed, whatever mechanism is used to create
the labels on GitHub needs to know about the old name of the label to
rename it. Otherwise, a new label would be created instead of renaming
the old one. This is supported via the optional `aliases` field on
labels, which accepts a list of alternative names.