Fix version comparison in GitHub Actions workflow  (#33)
The GitHub Actions workflow that runs the action for Labelflair on this
repository performs a version comparison to avoid running on release
PRs. This is necessary, since the default version in the action won't
be available as a release yet.

The command that was used to extract the version from the `action.yml`
did not work, which has prevented the check from running at all. This is
now fixed.