Separate testing GitHub Action from running it (#35)
We ran into issues where the test for the GitHub Action failed due to
new features in the local configuration file. To ensure that we both
execute the action and test new features as early in the development
cycle as possible, we have decided to split the action into two:

  - one workflow executes the GitHub Action in the repository
  - another workflow builds and runs the latest version of Labelflair

This provides us both with confidence that the GitHub Action works, and
allows us to run the latest development version of Labelflair against
this repository.