Support individual labels outside groups (#36)
The configuration has been extended to allow individual labels to be
defined outside of groups. Each individual label must have a `name` and
`color`, and can optionally have `description` and a list of `aliases`.
This feature is useful for one-off labels that have no related labels,
or for labels that should have a very specific color.